third_party/blink/web_tests/platform/win/external/wpt/html/semantics/forms/the-selectlist-element/selectlist-popover-position-with-zoom.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The popover should be top left positioned
  assert_equals: expected 0 but got 567
[FAIL] The popover should be top right positioned
  assert_equals: expected 0 but got 257
Harness: the test ran to completion.
